+++ CHANGELOG.md
@@ -7,8 +7,11 @@ Le format est basé sur [Tenez un Changelog](https://keepachangelog.com/fr/1.0.0
 ## [À venir dans la prochaine version]
 ### Ajouté
 - Visualiser le fait marquant avant de le sauvegarder
+- Système de rappel pour les événements : Notification, SMS et slack
 ### Modifié
 ### Réparé
+- L'accès de l'ADMIN à l'export d'un projet
+- Export feuilles de temps